pyproject.toml
[tool.poetry]
name = "easytask"
version = "0.1.0"
description = ""
authors = ["kirath2205 <singhkirath@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
mongoengine = "^0.29.1"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.54"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
